functions/requirements.txt
firebase_functions~=0.1.0
firebase_admin>=6.6.0
openai>=1.61.0
python-dotenv>=1.0.1